Move LLVM-dependent functions in CppUtil.h to LLVMUtil.h
