+++ .clang-format
@@ -2,4 +2,4 @@ BasedOnStyle: Microsoft
 PointerAlignment: Left
 IndentPPDirectives: AfterHash
 AllowShortFunctionsOnASingleLine: Empty
-ColumnLimit: 120
+ColumnLimit: 80
