+++ include/SVF-LLVM/DCHG.h
@@ -16,7 +16,8 @@
 
 #include "Graphs/GenericGraph.h"
 #include "Graphs/CHG.h"
-#include "SVF-LLVM/CPPUtil.h"
+#include "Util/CppUtil.h"
+#include "SVF-LLVM/BasicTypes.h"
 #include "SVFIR/SVFModule.h"
 #include "Util/SVFUtil.h"
 #include "Util/WorkList.h"
