+++ include/Util/config.h
@@ -1,0 +1,6 @@
+#ifndef CONFIG_H_IN
+#define CONFIG_H_IN
+
+#define PROJECT_PATH "/home/xudong/research/svf-whst"
+
+#endif
