+++ lib/SABER/SaberCondAllocator.cpp
@@ -34,7 +34,7 @@
 #include "Graphs/SVFG.h"
 #include <climits>
 #include <cmath>
-#include "SVF-LLVM/BasicTypes.h"//WHSTODO: remove
+#include "SVF-LLVM/BasicTypes.h" // TODO: Remove LLVM Header
 
 using namespace SVF;
 using namespace SVFUtil;
